Split the readers into 2 files
