+++ benchmarks/concurrent_read/readers/vendor.go
@@ -22,43 +22,9 @@ import (
 	"net/http"
 	"net/url"
 
-	"cloud.google.com/go/storage"
 	"github.com/jacobsa/gcloud/gcs"
 	"golang.org/x/oauth2/google"
 )
-
-type Client interface {
-	NewReader(objectName string) (io.ReadCloser, error)
-}
-
-func NewClient(ctx context.Context, protocol string, connections int, implementation string, bucketName string) (Client, error) {
-	transport := getTransport(protocol, connections)
-	switch implementation {
-	case "vendor":
-		return newVendorClient(ctx, transport, bucketName)
-	case "google":
-		return newGoogleClient(ctx, transport, bucketName)
-	default:
-		panic(fmt.Errorf("Unknown reader implementation: %q", implementation))
-	}
-}
-
-func getTransport(protocol string, connections int) (transport *http.Transport) {
-	switch protocol {
-	case "HTTP/1.1":
-		return &http.Transport{
-			MaxConnsPerHost: connections,
-			// This disables HTTP/2 in the transport.
-			TLSNextProto: make(
-				map[string]func(string, *tls.Conn) http.RoundTripper,
-			),
-		}
-	case "HTTP/2":
-		return http.DefaultTransport.(*http.Transport).Clone()
-	default:
-		panic(fmt.Errorf("Unknown tranport type: %q", protocol))
-	}
-}
 
 // Vendor reader depends on "github.com/jacobsa/gcloud/gcs"
 type vendorClient struct {
@@ -66,7 +32,7 @@ type vendorClient struct {
 	bucket gcs.Bucket
 }
 
-func newVendorClient(ctx context.Context, transport *http.Transport, bucketName string) (Client, error) {
+func NewVendorClient(ctx context.Context, protocol string, connections int, bucketName string) (*vendorClient, error) {
 	tokenSrc, err := google.DefaultTokenSource(
 		ctx,
 		gcs.Scope_FullControl,
@@ -79,7 +45,7 @@ func newVendorClient(ctx context.Context, transport *http.Transport, bucketName 
 		Url:         endpoint,
 		TokenSource: tokenSrc,
 		UserAgent:   "gcsfuse/dev Benchmark (concurrent_read)",
-		Transport:   transport,
+		Transport:   getTransport(protocol, connections),
 	}
 	conn, err := gcs.NewConn(config)
 	if err != nil {
@@ -106,21 +72,19 @@ func (c *vendorClient) NewReader(objectName string) (io.ReadCloser, error) {
 	)
 }
 
-// Official reader depends on "cloud.google.com/go/storage"
-type googleClient struct {
-	ctx    context.Context
-	bucket *storage.BucketHandle
-}
-
-func newGoogleClient(ctx context.Context, transport *http.Transport, bucketName string) (Client, error) {
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		return nil, err
+func getTransport(protocol string, connections int) (transport *http.Transport) {
+	switch protocol {
+	case "HTTP/1.1":
+		return &http.Transport{
+			MaxConnsPerHost: connections,
+			// This disables HTTP/2 in the transport.
+			TLSNextProto: make(
+				map[string]func(string, *tls.Conn) http.RoundTripper,
+			),
+		}
+	case "HTTP/2":
+		return http.DefaultTransport.(*http.Transport).Clone()
+	default:
+		panic(fmt.Errorf("Unsupported protocol: %q", protocol))
 	}
-	bucket := client.Bucket(bucketName)
-	return &googleClient{ctx, bucket}, nil
-}
-
-func (c *googleClient) NewReader(objectName string) (io.ReadCloser, error) {
-	return c.bucket.Object(objectName).NewReader(c.ctx)
 }
